Rename lib/types.py to not shadow Python module
