+++ extra_game_handlers.py
@@ -1,6 +1,6 @@
 """Functions for the user to implement when the config file is not adequate to express bot requirements."""
 from lib import model
-from lib.types import OPTIONS_TYPE
+from lib.lichess_types import OPTIONS_TYPE
 
 
 def game_specific_options(game: model.Game) -> OPTIONS_TYPE:  # noqa: ARG001
