+++ homemade.py
@@ -7,7 +7,7 @@ import chess
 from chess.engine import PlayResult, Limit
 import random
 from lib.engine_wrapper import MinimalEngine
-from lib.types import MOVE, HOMEMADE_ARGS_TYPE
+from lib.lichess_types import MOVE, HOMEMADE_ARGS_TYPE
 import logging
