+++ lib/lichess_types.py
