+++ lib/model.py
@@ -7,7 +7,7 @@ from enum import Enum
 from lib.timer import Timer, msec, seconds, sec_str, to_msec, to_seconds, years
 from lib.config import Configuration
 from collections import defaultdict, Counter
-from lib.types import UserProfileType, ChallengeType, GameEventType, PlayerType
+from lib.lichess_types import UserProfileType, ChallengeType, GameEventType, PlayerType
 
 logger = logging.getLogger(__name__)
